Include SD card type requirements
I was recently gifted a Raspberry Pi 2 and in purchasing an SD card, I tried to work out what type of SD card to get while at the store. This SD Cards page was the main result, and implied that it required a regular SD card.

Then I got home and found it was MicroSD card slot. So I thought I'd update the documentation so it's a bit more clear about what type of SD card should be purchased and hopefully others won't do my mistake!